working on trsansferring modal over
