+++ client/src/components/contactBar.js
@@ -1,145 +1,6 @@
 import React, { Component } from 'react';
-import Modal from 'react-modal';
-import Alert from 'react-s-alert';
-import 'react-s-alert/dist/s-alert-css-effects/slide.css';
 
 class ContactBar extends Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      modalIsOpen: false,
-      email: "your email",
-      subject: "subject",
-      body: "your email",
-      emailValidated: false
-    }
-    this.handleSubmit = this.handleSubmit.bind(this);
-    this.handleInputChange = this.handleInputChange.bind(this);
-    this.customStyles = {
-      overlay : {
-        position          : 'fixed',
-        top               : 0,
-        left              : 0,
-        right             : 0,
-        bottom            : 0,
-        backgroundColor   : 'rgba(0, 0, 0, 0.75)'
-      },
-      content : {
-        position                   : 'absolute',
-        top                        : '50%',
-        left                       : '50%',
-        right                      : 'auto',
-        bottom                     : 'auto',
-        border                     : '1px solid #ccc',
-        background                 : '#fff',
-        overflow                   : 'auto',
-        WebkitOverflowScrolling    : 'touch',
-        borderRadius               : '4px',
-        outline                    : 'none',
-        padding                    : '20px',
-        width                      : '60%',
-        height                     : '85%',
-        marginRight                : '-50%',
-        transform                  : 'translate(-50%, -50%)'
-
-      }
-
-    };
-    this.openModal = this.openModal.bind(this);
-    this.afterOpenModal = this.afterOpenModal.bind(this);
-    this.closeModal = this.closeModal.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-  }
-  openModal() {
-    this.setState({modalIsOpen: true});
-  }
-
-  afterOpenModal() {
-    // references are now sync'd and can be accessed.
-    this.subtitle.style.color = 'black';
-  }
-
-  closeModal() {
-    this.setState({modalIsOpen: false});
-  }
-  handleInputChange (event) {
-    event.preventDefault();
-    const target = event.target;
-    const value = target.value;
-    const name = target.name;
-    this.setState({
-      [name]: value
-    });
-  }
-  validateEmail(value) {
-    var re = /^(([^<>()[\]\\.,;:\s@\"]+(\.[^<>()[\]\\.,;:\s@\"]+)*)|(\".+\"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
-    return re.test(value);
-  }
-  handleSubmit(event) {
-    const self = this;
-    event.preventDefault();
-    if (this.validateEmail(this.state.email)) {
-      this.setState({emailValidated: true});
-      var formData = {
-        "email": this.state.email,
-        "body": this.state.body,
-        "subject": this.state.subject
-      };
-      fetch('api/contact', {
-      method: 'post',
-      headers: {
-        "Content-type": "application/json; charset=UTF-8"
-      },
-        "body": JSON.stringify(formData)
-    })
-    .then(function (data) {
-      console.log('Request succeeded with JSON response', data);
-      if (data.status === 500) {
-        Alert.error('<h5 class="center">There was a fatal error. Your email could not be sent. Please try again.</h5>', {
-              position: 'bottom',
-              effect: 'slide',
-              html: true,
-              beep: true,
-              timeout: 3000,
-              offset: 100
-          });
-      }
-      else {
-        self.closeModal();
-        self.setState({email: "your email",
-        subject: "subject",
-        body: "your email"});
-        Alert.success('<h4 class="center">Your email was sent. Thank You! I will answer shortly.</h4>', {
-              position: 'top',
-              effect: 'slide',
-              html: true,
-              beep: true,
-              timeout: 2000,
-              offset: 100
-          });
-      }
-
-    })
-    .catch(function (error) {
-      console.log('Request failed', error);
-
-    });
-  }
-    else {
-      Alert.warning('<h5>That is not a valid email...</h5>', {
-            position: 'top-left',
-            effect: 'slide',
-            html: true,
-            beep: true,
-            timeout: 3000,
-            offset: 100
-        });
-    }
-
-  }
-
-
   render () {
     return (
       <div className="col-xs-3 col-sm-2 col-md-1 col-lg-1" id="contactBar">
@@ -164,63 +25,6 @@ class ContactBar extends Component {
           <div className="col-xs-12 col-sm-12 col-md-12 col-lg-12 pull-left">
             <img onClick={this.openModal} style={{cursor:'pointer'}} className="img-responsive" src="./images/mail.png" alt="Email Me" />
           </div>
-          <div>
-            <Modal
-              isOpen={this.state.modalIsOpen}
-              onAfterOpen={this.afterOpenModal}
-              onRequestClose={this.closeModal}
-              style={this.customStyles}
-              contentLabel="Email Modal"
-              >
-                <div className='col-lg-12' id='emailForm'>
-                <h2 ref={subtitle => this.subtitle = subtitle}>Send me an Email</h2>
-                <div className='col-lg-8 col-lg-offset-2'>
-
-                  <form action="" onSubmit={this.handleSubmit}>
-                    <br/>
-                    <label htmlFor="emailLine">
-                      Your Email Address:
-                      </label>
-                      <input
-                        style={{width: 400 }}
-                        id='emailLine'
-                        name="email"
-                        type="text"
-                        value={this.state.email}
-                        onChange={this.handleInputChange} />
-
-                      <br/>
-                      <label htmlFor="subjectLine">
-                        Subject:
-                        </label>
-                        <input
-                          style={{width: 400 }}
-                          id='subjectLine'
-                          name="subject"
-                          type="text"
-                          value={this.state.subject}
-                          onChange={this.handleInputChange} />
-
-                        <br/>
-                        <label htmlFor="emailBody">
-                          Email Body:
-                          </label>
-                          <textarea
-                            style={{ height: 150, width: 400 }}
-                            id='emailBody'
-                            name="body"
-                            type="text"
-                            value={this.state.body}
-                            onChange={this.handleInputChange} />
-
-                        <button className='btn btn-warning' onClick={this.closeModal}>close</button>
-                        <button type='submit' className='btn btn-success' >Submit</button>
-                      </form>
-
-                    </div>
-                  </div>
-                  </Modal>
-                </div>
               </div>
             );
           }
